+++ core/src/main/java/org/ehcache/spi/Ehcaching.java
@@ -23,5 +23,5 @@ import org.ehcache.config.Configuration;
  * @author Alex Snaps
  */
 public interface Ehcaching {
-  PersistentCacheManager createCacheManager(Configuration configuration, ServiceProvider serviceProvider);
+  PersistentCacheManager createCacheManager(Configuration configuration, ServiceLocator serviceLocator);
 }
