+++ core/src/test/java/org/ehcache/StandaloneCacheBuilderTest.java
@@ -17,15 +17,13 @@
 package org.ehcache;
 
 import org.ehcache.config.StandaloneCacheConfiguration;
-import org.ehcache.spi.ServiceProvider;
+import org.ehcache.spi.ServiceLocator;
 import org.junit.Test;
 
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
-import static org.ehcache.StandaloneCacheBuilder.newCacheBuilder;
-import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.Assert.assertThat;
@@ -40,7 +38,7 @@ public class StandaloneCacheBuilderTest {
       public StandaloneCacheBuilder<String, Object, TestStandaloneCache<String, Object>> builder(final StandaloneCacheBuilder<String, Object, ? extends StandaloneCache<String, Object>> builder) {
         return new StandaloneCacheBuilder<String, Object, TestStandaloneCache<String, Object>>(String.class, Object.class) {
           @Override
-          TestStandaloneCache<String, Object> build(final ServiceProvider serviceProvider) {
+          TestStandaloneCache<String, Object> build(final ServiceLocator serviceProvider) {
             return new TestStandaloneCache<String, Object>();
           }
         };
