+++ impl/src/test/java/org/ehcache/internal/CacheTest.java
@@ -19,11 +19,11 @@ package org.ehcache.internal;
 import java.io.Serializable;
 
 import org.ehcache.Cache;
+import org.ehcache.spi.ServiceLocator;
 import org.junit.Test;
 import org.ehcache.internal.cachingtier.TieredCache;
 import org.ehcache.exceptions.CacheAccessException;
 import org.ehcache.internal.serialization.JavaSerializationProvider;
-import org.ehcache.spi.ServiceProvider;
 import org.ehcache.internal.serialization.SerializationProvider;
 
 import static org.hamcrest.core.IsNot.*;
@@ -42,7 +42,7 @@ public class CacheTest {
     final HeapResource heapResource = new HeapResource(null);
     final Cache<String, String> cachingTier = heapResource.createCache(String.class, String.class);
     Cache<String, String> cache = new TieredCache<String, String>(cachingTier,
-        String.class, String.class, new ServiceProvider(heapResource), new HeapResourceCacheConfiguration(1000));
+        String.class, String.class, new ServiceLocator(heapResource), new HeapResourceCacheConfiguration(1000));
     cache.put("foo", "bar");
     assertThat(cache, hasEntry("foo", "bar"));
   }
@@ -50,11 +50,11 @@ public class CacheTest {
   @Test
   public void testSerializingCache() throws CacheAccessException {
     final SerializationProvider serialization = new JavaSerializationProvider();
-    final ServiceProvider serviceProvider = new ServiceProvider(serialization);
-    final SerializingResource serializingResource = new SerializingResource(serviceProvider);
-    final HeapResource heapResource = new HeapResource(serviceProvider);
+    final ServiceLocator serviceLocator = new ServiceLocator(serialization);
+    final SerializingResource serializingResource = new SerializingResource(serviceLocator);
+    final HeapResource heapResource = new HeapResource(serviceLocator);
     final Cache<String, String> cachingTier = serializingResource.createCache(String.class, String.class);
-    Cache<String, String> cache = new TieredCache<String, String>(cachingTier, String.class, String.class, new ServiceProvider(heapResource), new HeapResourceCacheConfiguration(1000));
+    Cache<String, String> cache = new TieredCache<String, String>(cachingTier, String.class, String.class, new ServiceLocator(heapResource), new HeapResourceCacheConfiguration(1000));
     cache.put("foo", "bar");
     assertThat(cache, hasEntry("foo", "bar"));
   }
@@ -62,11 +62,11 @@ public class CacheTest {
   @Test
   public void testSerializingCacheWithNonSerializableType() throws CacheAccessException {
     final SerializationProvider serialization = new JavaSerializationProvider();
-    final ServiceProvider serviceProvider = new ServiceProvider(serialization);
-    final SerializingResource serializingResource = new SerializingResource(serviceProvider);
-    final HeapResource heapResource = new HeapResource(serviceProvider);
+    final ServiceLocator serviceLocator = new ServiceLocator(serialization);
+    final SerializingResource serializingResource = new SerializingResource(serviceLocator);
+    final HeapResource heapResource = new HeapResource(serviceLocator);
     final Cache<String, Serializable> cachingTier = serializingResource.createCache(String.class, Serializable.class);
-    Cache<String, Serializable> cache = new TieredCache<String, Serializable>(cachingTier, String.class, Serializable.class, new ServiceProvider(heapResource), new HeapResourceCacheConfiguration(1000));
+    Cache<String, Serializable> cache = new TieredCache<String, Serializable>(cachingTier, String.class, Serializable.class, new ServiceLocator(heapResource), new HeapResourceCacheConfiguration(1000));
     try {
       cache.put("foo", new Object[] { new Object() });
       fail();
